+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changed
 
 - Make error level default in Log handler
+- Do not export `LogHandler`
 
 
 ## 0.2.2 - 2017-07-07